stained-mask/requirements.txt
functions-framework>=3.0.0
google-cloud-storage
google-cloud-firestore
Pillow